requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dash==4.0.0
dash-bootstrap-components==2.0.4
espn-api==0.45.1
Flask==3.1.3
fonttools==4.61.1
gunicorn==25.1.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.17.0
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plotly==6.5.2
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.32.5
retrying==1.4.2
setuptools==82.0.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.2.3
Werkzeug==3.1.6
zipp==3.23.0
